fix arena cgo windows
